# frontend/openapi.json
{"openapi":"3.0.2","info":{"title":"RentAndYou API","version":"0.1.0"},"paths":{"/api/v1/vehicles/":{"get":{"tags":["Vehicle"],"summary":"Get Vehicles","operationId":"get_vehicles_api_v1_vehicles__get","parameters":[{"required":false,"schema":{"title":"Office Id","type":"integer"},"name":"office_id","in":"query"},{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Vehicles Api V1 Vehicles  Get","type":"array","items":{"$ref":"#/components/schemas/VehicleReadModel"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Vehicle"],"summary":"Create Vehicle","operationId":"create_vehicle_api_v1_vehicles__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleCreateModel"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleReadModel"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageVehicleAlreadyExists"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicles/{id_}/":{"get":{"tags":["Vehicle"],"summary":"Get Vehicle","operationId":"get_vehicle_api_v1_vehicles__id____get","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageVehicleNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Vehicle"],"summary":"Delete Vehicle","operationId":"delete_vehicle_api_v1_vehicles__id____delete","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageVehicleNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Vehicle"],"summary":"Update Vehicle","operationId":"update_vehicle_api_v1_vehicles__id____patch","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleUpdateModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageVehicleNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rents/":{"get":{"tags":["Rent"],"summary":"Get Rents","operationId":"get_rents_api_v1_rents__get","parameters":[{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Rents Api V1 Rents  Get","type":"array","items":{"$ref":"#/components/schemas/RentReadModel"}}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageRentNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Rent"],"summary":"Create Rent","operationId":"create_rent_api_v1_rents__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentCreateModel"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentReadModel"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageRentAlreadyExists"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rents/{id_}/":{"get":{"tags":["Rent"],"summary":"Get Rent","operationId":"get_rent_api_v1_rents__id____get","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageRentNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Rent"],"summary":"Delete Rent","operationId":"delete_rent_api_v1_rents__id____delete","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageRentNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Rent"],"summary":"Update Rent","operationId":"update_rent_api_v1_rents__id____patch","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentUpdateModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageRentNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/offices/":{"get":{"tags":["Office"],"summary":"Get Offices","operationId":"get_offices_api_v1_offices__get","parameters":[{"required":false,"schema":{"title":"City","type":"string"},"name":"city","in":"query"},{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Offices Api V1 Offices  Get","type":"array","items":{"$ref":"#/components/schemas/OfficeReadModel"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Office"],"summary":"Create Office","operationId":"create_office_api_v1_offices__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeCreateModel"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeReadModel"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageOfficeAlreadyExists"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/offices/{id_}/":{"get":{"tags":["Office"],"summary":"Get Office","operationId":"get_office_api_v1_offices__id____get","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageOfficeNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Office"],"summary":"Delete Office","operationId":"delete_office_api_v1_offices__id____delete","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageOfficeNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Office"],"summary":"Update Office","operationId":"update_office_api_v1_offices__id____patch","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeUpdateModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageOfficeNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/offices/{office_id}/vehicles/available/":{"get":{"tags":["Office"],"summary":"Get Office Vehicles Available","operationId":"get_office_vehicles_available_api_v1_offices__office_id__vehicles_available__get","parameters":[{"required":true,"schema":{"title":"Office Id","type":"integer"},"name":"office_id","in":"path"},{"required":false,"schema":{"title":"Pickup Date","type":"string","format":"date-time"},"name":"pickup_date","in":"query"},{"required":false,"schema":{"title":"Return Date","type":"string","format":"date-time"},"name":"return_date","in":"query"},{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Office Vehicles Available Api V1 Offices  Office Id  Vehicles Available  Get","type":"array","items":{"$ref":"#/components/schemas/VehicleReadModel"}}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageOfficeNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/":{"get":{"tags":["User"],"summary":"Get Users","operationId":"get_users_api_v1_users__get","parameters":[{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Users Api V1 Users  Get","type":"array","items":{"$ref":"#/components/schemas/UserReadModel"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["User"],"summary":"Create User","operationId":"create_user_api_v1_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateModel"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReadModel"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageUserAlreadyExists"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/{id_}/":{"get":{"tags":["User"],"summary":"Get User","operationId":"get_user_api_v1_users__id____get","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageUserNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["User"],"summary":"Delete User","operationId":"delete_user_api_v1_users__id____delete","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageUserNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["User"],"summary":"Update User","operationId":"update_user_api_v1_users__id____patch","parameters":[{"required":true,"schema":{"title":"Id ","type":"integer"},"name":"id_","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReadModel"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageUserNotFound"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BodyType":{"title":"BodyType","enum":["SUV","Sedan","Hatchback","Other"],"type":"string","description":"An enumeration."},"Brand":{"title":"Brand","enum":["Mercedes","Audi","Seat","Other"],"type":"string","description":"An enumeration."},"ErrorMessageOfficeAlreadyExists":{"title":"ErrorMessageOfficeAlreadyExists","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Office already exists"}}},"ErrorMessageOfficeNotFound":{"title":"ErrorMessageOfficeNotFound","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Office does not exist."}}},"ErrorMessageRentAlreadyExists":{"title":"ErrorMessageRentAlreadyExists","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Rent already exists"}}},"ErrorMessageRentNotFound":{"title":"ErrorMessageRentNotFound","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Rent does not exist."}}},"ErrorMessageUserAlreadyExists":{"title":"ErrorMessageUserAlreadyExists","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"User already exists"}}},"ErrorMessageUserNotFound":{"title":"ErrorMessageUserNotFound","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"User does not exist."}}},"ErrorMessageVehicleAlreadyExists":{"title":"ErrorMessageVehicleAlreadyExists","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Vehicle already exists"}}},"ErrorMessageVehicleNotFound":{"title":"ErrorMessageVehicleNotFound","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string","example":"Vehicle does not exist."}}},"Fares":{"title":"Fares","enum":["Smart","Plus","Premium"],"type":"string","description":"An enumeration."},"GearBox":{"title":"GearBox","enum":["Manual","Automatic"],"type":"string","description":"An enumeration."},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"OfficeCreateModel":{"title":"OfficeCreateModel","required":["name","phone","address","postal_code","city","email"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"phone":{"title":"Phone","type":"string"},"address":{"title":"Address","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"city":{"title":"City","type":"string"},"email":{"title":"Email","type":"string","example":"test@test.com"}},"description":"OfficeCreateModel represents a write model to create a office"},"OfficeReadModel":{"title":"OfficeReadModel","required":["name","phone","address","postal_code","city","email","id_","is_active","is_deleted","created_at","updated_at"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"phone":{"title":"Phone","type":"string"},"address":{"title":"Address","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"city":{"title":"City","type":"string"},"email":{"title":"Email","type":"string","example":"test@test.com"},"id_":{"title":"Id ","type":"integer"},"is_active":{"title":"Is Active","type":"boolean","example":true},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"},"rents":{"title":"Rents","type":"array","items":{"type":"integer"}}},"description":"OfficeReadModel represents data structure as a read model"},"OfficeUpdateModel":{"title":"OfficeUpdateModel","type":"object","properties":{"name":{"title":"Name","type":"string"},"phone":{"title":"Phone","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"email":{"title":"Email","type":"string"},"is_active":{"title":"Is Active","type":"boolean","example":true},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true}},"description":"OfficeUpdateModel represents a write model to update a office"},"RentCreateModel":{"title":"RentCreateModel","required":["ammount","total_days","pickup_date","return_date","office_id","vehicle_id","user_id"],"type":"object","properties":{"ammount":{"title":"Ammount","type":"number"},"total_days":{"title":"Total Days","type":"integer"},"pickup_date":{"title":"Pickup Date","type":"string","format":"date-time"},"return_date":{"title":"Return Date","type":"string","format":"date-time"},"office_id":{"title":"Office Id","type":"integer"},"vehicle_id":{"title":"Vehicle Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"}},"description":"RentCreateModel represents a write model to create a rent"},"RentReadModel":{"title":"RentReadModel","required":["ammount","total_days","pickup_date","return_date","office_id","vehicle_id","user_id","id_","is_deleted","created_at","updated_at"],"type":"object","properties":{"ammount":{"title":"Ammount","type":"number"},"total_days":{"title":"Total Days","type":"integer"},"pickup_date":{"title":"Pickup Date","type":"string","format":"date-time"},"return_date":{"title":"Return Date","type":"string","format":"date-time"},"office_id":{"title":"Office Id","type":"integer"},"vehicle_id":{"title":"Vehicle Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"id_":{"title":"Id ","type":"integer"},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"}},"description":"RentReadModel represents data structure as a read model"},"RentUpdateModel":{"title":"RentUpdateModel","required":["id_"],"type":"object","properties":{"id_":{"title":"Id ","type":"integer"},"ammount":{"title":"Ammount","type":"number","example":100.0},"total_days":{"title":"Total Days","type":"integer","example":5},"is_active":{"title":"Is Active","type":"boolean","example":true},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true}},"description":"RentUpdateModel represents a write model to update a rent"},"UserCreateModel":{"title":"UserCreateModel","required":["document_type","document_id","first_name","postal_code","address","city","email"],"type":"object","properties":{"document_type":{"title":"Document Type","type":"string"},"document_id":{"title":"Document Id","type":"string"},"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"hashed_password":{"title":"Hashed Password","type":"string"},"email":{"title":"Email","type":"string","example":"test@test.com"}},"description":"UserCreateModel represents a write model to create a user"},"UserReadModel":{"title":"UserReadModel","required":["is_active","is_deleted","created_at","updated_at"],"type":"object","properties":{"document_type":{"title":"Document Type","type":"string"},"document_id":{"title":"Document Id","type":"string"},"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"hashed_password":{"title":"Hashed Password","type":"string"},"email":{"title":"Email","type":"string","example":"test@test.com"},"id_":{"title":"Id ","type":"integer"},"is_active":{"title":"Is Active","type":"boolean","example":true},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"},"rents":{"title":"Rents","type":"array","items":{"type":"integer"}}},"description":"UserReadModel represents data structure as a read model"},"UserUpdateModel":{"title":"UserUpdateModel","required":["email"],"type":"object","properties":{"document_type":{"title":"Document Type","type":"string"},"document_id":{"title":"Document Id","type":"string"},"first_name":{"title":"First Name","type":"string"},"last_name":{"title":"Last Name","type":"string"},"postal_code":{"title":"Postal Code","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"hashed_password":{"title":"Hashed Password","type":"string"},"email":{"title":"Email","type":"string","example":"test@test.com"},"is_active":{"title":"Is Active","type":"boolean","example":true},"is_deleted":{"title":"Is Deleted","type":"boolean","example":true}},"description":"UserUpdateModel represents a write model to update a user"},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"VehicleCreateModel":{"title":"VehicleCreateModel","required":["model","version","color","brand","kms","license_plate","purchase_date","gearbox","body_type","price_per_day","passengers","office_id"],"type":"object","properties":{"model":{"title":"Model","type":"string"},"version":{"title":"Version","type":"string"},"color":{"title":"Color","type":"string"},"brand":{"$ref":"#/components/schemas/Brand"},"kms":{"title":"Kms","type":"integer"},"license_plate":{"title":"License Plate","type":"string"},"purchase_date":{"title":"Purchase Date","type":"string","format":"date"},"gearbox":{"$ref":"#/components/schemas/GearBox"},"body_type":{"$ref":"#/components/schemas/BodyType"},"price_per_day":{"title":"Price Per Day","type":"number"},"passengers":{"title":"Passengers","type":"integer"},"avg_consumption":{"title":"Avg Consumption","type":"number"},"fare":{"$ref":"#/components/schemas/Fares"},"image_url":{"title":"Image Url","type":"string"},"office_id":{"title":"Office Id","type":"integer"}}},"VehicleReadModel":{"title":"VehicleReadModel","required":["id_","created_at","updated_at"],"type":"object","properties":{"model":{"title":"Model","type":"string"},"version":{"title":"Version","type":"string"},"color":{"title":"Color","type":"string"},"brand":{"$ref":"#/components/schemas/Brand"},"kms":{"title":"Kms","type":"integer"},"license_plate":{"title":"License Plate","type":"string"},"purchase_date":{"title":"Purchase Date","type":"string","format":"date"},"gearbox":{"$ref":"#/components/schemas/GearBox"},"body_type":{"$ref":"#/components/schemas/BodyType"},"price_per_day":{"title":"Price Per Day","type":"number"},"passengers":{"title":"Passengers","type":"integer"},"avg_consumption":{"title":"Avg Consumption","type":"number"},"fare":{"$ref":"#/components/schemas/Fares"},"image_url":{"title":"Image Url","type":"string"},"office_id":{"title":"Office Id","type":"integer"},"id_":{"title":"Id ","type":"integer"},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"}}},"VehicleUpdateModel":{"title":"VehicleUpdateModel","type":"object","properties":{"model":{"title":"Model","type":"string"},"version":{"title":"Version","type":"string"},"color":{"title":"Color","type":"string"},"brand":{"$ref":"#/components/schemas/Brand"},"kms":{"title":"Kms","type":"integer"},"license_plate":{"title":"License Plate","type":"string"},"purchase_date":{"title":"Purchase Date","type":"string","format":"date"},"gearbox":{"$ref":"#/components/schemas/GearBox"},"body_type":{"$ref":"#/components/schemas/BodyType"},"price_per_day":{"title":"Price Per Day","type":"number"},"passengers":{"title":"Passengers","type":"integer"},"avg_consumption":{"title":"Avg Consumption","type":"number"},"fare":{"$ref":"#/components/schemas/Fares"},"image_url":{"title":"Image Url","type":"string"},"office_id":{"title":"Office Id","type":"integer"},"is_deleted":{"title":"Is Deleted","type":"boolean","default":false,"example":true}}}}}}